requirements.txt.txt
# Core dependencies
telebot==0.0.5
yfinance==0.2.28
pandas==2.0.3
numpy==1.24.3
requests==2.31.0
flask==2.3.2

# AI providers
groq==0.4.2
google-generativeai==0.3.2

# Database
sqlite3  # Built-in

# Utilities
psutil==5.9.5
pytz==2023.3
python-dotenv==1.0.0

# News (optional)
finnhub-python==2.4.18

# For better performance
cachetools==5.3.1